+++ src/utils/utils.cpp
@@ -623,7 +623,7 @@ bool mp::Utils::invalid_target_path(const QString& target_path) const
     return matcher.match(target_path).hasMatch();
 }
 
-int mp::Utils::random_int(int a, int b) const
+intmax_t mp::Utils::random_int(intmax_t a, intmax_t b) const
 {
     static std::default_random_engine gen = [] {
         // seed the rng with the time at first call
@@ -634,7 +634,7 @@ int mp::Utils::random_int(int a, int b) const
     if (a > b) // avoid undefined behavior, this can only happen by programmer error
         throw std::logic_error(fmt::format("random range [{}, {}] is invalid", a, b));
 
-    std::uniform_int_distribution<int> dist{a, b};
+    std::uniform_int_distribution<intmax_t> dist{a, b};
 
     return dist(gen);
 }
